% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lead_data_loaders.R
\name{aggregate_cohort_data}
\alias{aggregate_cohort_data}
\title{Aggregate cohort data by census tract and income bracket}
\usage{
aggregate_cohort_data(data, dataset, vintage, verbose = FALSE)
}
\description{
Aggregate cohort data by census tract and income bracket
}
\keyword{internal}
