% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-helpers.R
\name{db_exists}
\alias{db_exists}
\title{Check if Database Exists}
\usage{
db_exists(test = FALSE)
}
\arguments{
\item{test}{Logical, check test database instead of production}
}
\value{
Logical, TRUE if database exists
}
\description{
Check if Database Exists
}
\keyword{internal}
