% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lead_data_loaders.R
\name{download_and_merge_states}
\alias{download_and_merge_states}
\title{Download and merge data from multiple states}
\usage{
download_and_merge_states(dataset, vintage, states, verbose = TRUE)
}
\arguments{
\item{dataset}{Character, "ami" or "fpl"}

\item{vintage}{Character, "2018" or "2022"}

\item{states}{Character vector of state abbreviations}

\item{verbose}{Logical, print progress messages}
}
\value{
Combined tibble with data from all states
}
\description{
Download and merge data from multiple states
}
\keyword{internal}
