% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lead_data_loaders.R
\name{download_census_tract_data}
\alias{download_census_tract_data}
\title{Download census tract data from OpenEI}
\usage{
download_census_tract_data(verbose = FALSE)
}
\description{
Download census tract data from OpenEI
}
\keyword{internal}
