% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lead_data_loaders.R
\name{download_lead_data}
\alias{download_lead_data}
\title{Download LEAD data from OpenEI}
\usage{
download_lead_data(dataset, vintage, states = NULL, verbose = FALSE)
}
\description{
Download LEAD data from OpenEI
}
\keyword{internal}
