% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lead_data_loaders.R
\name{get_all_states}
\alias{get_all_states}
\title{Get all state abbreviations}
\usage{
get_all_states()
}
\value{
Character vector of all 51 state abbreviations (50 states + DC)
}
\description{
Get all state abbreviations
}
\keyword{internal}
