% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_utils.R, R/lead_data_loaders.R
\name{get_cache_dir}
\alias{get_cache_dir}
\title{Get the emburden cache directory}
\usage{
get_cache_dir()

get_cache_dir()
}
\description{
Get the emburden cache directory

Get cache directory for downloaded files
}
\keyword{internal}
