% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_utils.R
\name{get_database_path}
\alias{get_database_path}
\title{Get the full path to the emburden database file}
\usage{
get_database_path()
}
\description{
Get the full path to the emburden database file
}
\keyword{internal}
