% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_dataset_info}
\alias{get_dataset_info}
\title{Get Dataset Information}
\usage{
get_dataset_info()
}
\value{
Data frame with dataset information
}
\description{
Returns metadata about available LEAD datasets.
}
\examples{
get_dataset_info()
}
