% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-helpers.R
\name{get_db_connection}
\alias{get_db_connection}
\title{Get Database Connection}
\usage{
get_db_connection(test = FALSE)
}
\arguments{
\item{test}{Logical, connect to test database instead of production}
}
\value{
DBI connection object
}
\description{
Get Database Connection
}
\keyword{internal}
