% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lead_data_loaders.R
\name{get_state_fips}
\alias{get_state_fips}
\title{Get state FIPS codes from abbreviations}
\usage{
get_state_fips(state_abbrs)
}
\description{
Get state FIPS codes from abbreviations
}
\keyword{internal}
