% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenodo.R
\name{get_zenodo_config}
\alias{get_zenodo_config}
\title{Get Zenodo Record Information}
\usage{
get_zenodo_config()
}
\value{
List with Zenodo record information
}
\description{
Returns the Zenodo DOI and file information for emburden datasets.
}
\keyword{internal}
