% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{list_income_brackets}
\alias{list_income_brackets}
\title{List Available Income Brackets}
\usage{
list_income_brackets(dataset = c("ami", "fpl"), vintage = "2022")
}
\arguments{
\item{dataset}{Character, either "ami" or "fpl"}

\item{vintage}{Character, "2018" or "2022"}
}
\value{
Character vector of income bracket labels
}
\description{
Returns the income brackets available for a given dataset and vintage.
}
\examples{
list_income_brackets("ami", "2022")
list_income_brackets("fpl", "2018")
}
