% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_burden.R
\name{print.energy_burden_comparison}
\alias{print.energy_burden_comparison}
\title{Print Comparison Summary}
\usage{
\method{print}{energy_burden_comparison}(x, ...)
}
\arguments{
\item{x}{Comparison result from compare_energy_burden()}

\item{...}{Additional arguments (not used)}
}
\description{
Pretty-print a comparison table from compare_energy_burden()
}
