% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lead_data_loaders.R
\name{standardize_cohort_columns}
\alias{standardize_cohort_columns}
\title{Standardize cohort column names across vintages}
\usage{
standardize_cohort_columns(data, dataset, vintage)
}
\description{
Standardize cohort column names across vintages
}
\keyword{internal}
