% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{to_big}
\alias{to_big}
\title{Format Large Numbers with Thousand Separators}
\usage{
to_big(x)
}
\arguments{
\item{x}{Numeric vector to format}
}
\value{
Character vector of formatted numbers
}
\description{
Converts numeric values to formatted strings with thousand separators (commas).
}
\examples{
# Format large numbers
to_big(c(1000, 25000, 1000000))
}
