% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lead_data_loaders.R
\name{try_import_to_database}
\alias{try_import_to_database}
\title{Try to import cohort data to database}
\usage{
try_import_to_database(data, dataset, vintage, verbose = FALSE)
}
\description{
Try to import cohort data to database
}
\keyword{internal}
