% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lead_data_loaders.R
\name{try_import_tracts_to_database}
\alias{try_import_tracts_to_database}
\title{Try to import census tract data to database}
\usage{
try_import_tracts_to_database(data, verbose = FALSE)
}
\description{
Try to import census tract data to database
}
\keyword{internal}
