% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lead_data_loaders.R
\name{try_load_from_csv}
\alias{try_load_from_csv}
\title{Try to load cohort data from CSV}
\usage{
try_load_from_csv(dataset, vintage, verbose = FALSE)
}
\description{
Try to load cohort data from CSV
}
\keyword{internal}
