% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lead_data_loaders.R
\name{try_load_from_database}
\alias{try_load_from_database}
\title{Try to load cohort data from database}
\usage{
try_load_from_database(dataset, vintage, verbose = FALSE)
}
\description{
Try to load cohort data from database
}
\keyword{internal}
