% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lead_data_loaders.R
\name{try_load_tracts_from_database}
\alias{try_load_tracts_from_database}
\title{Try to load census tract data from database}
\usage{
try_load_tracts_from_database(verbose = FALSE)
}
\description{
Try to load census tract data from database
}
\keyword{internal}
