% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emburden-package.R
\docType{package}
\name{emburden-package}
\alias{emburden}
\alias{emburden-package}
\title{emburden: Energy Burden Analysis Using Net Energy Return Methodology}
\description{
Provides tools for calculating and analyzing household energy burden using the Net Energy Return aggregation methodology. Functions support weighted statistical calculations across geographic and demographic cohorts, with utilities for formatting results into publication-ready tables. Methods are based on Scheier & Kittner (2022) \doi{10.1038/s41467-021-27673-y}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ericscheier/emburden}
  \item \url{https://ericscheier.info/emburden/}
  \item Report bugs at \url{https://github.com/ericscheier/emburden/issues}
}

}
\author{
\strong{Maintainer}: Eric Scheier \email{eric@scheier.org} [copyright holder]

}
\keyword{internal}
